+++ src/main/java/icyllis/modernui/impl/setting/SettingCategoryGroup.java
@@ -23,7 +23,7 @@ import icyllis.modernui.font.FontTools;
 import icyllis.modernui.font.IFontRenderer;
 import icyllis.modernui.gui.scroll.ScrollWindow;
 import icyllis.modernui.gui.scroll.UniformScrollGroup;
-import icyllis.modernui.gui.math.Color3i;
+import icyllis.modernui.gui.math.Color3f;
 import net.minecraft.util.text.TextFormatting;
 
 import java.util.List;
@@ -60,7 +60,7 @@ public class SettingCategoryGroup extends UniformScrollGroup<SettingEntry> {
 
     @Override
     public void draw(float time) {
-        fontRenderer.drawString(title, centerX - 160, y1 + 14, Color3i.WHILE, 1.0f, TextAlign.LEFT);
+        fontRenderer.drawString(title, centerX - 160, y1 + 14, Color3f.WHILE, 1.0f, TextAlign.LEFT);
         super.draw(time);
     }
